feat: display node polyfill hint in error overlay
