+++ packages/core/src/helpers/stats.ts
@@ -5,87 +5,7 @@ import type { Rspack } from '../types';
 import { isMultiCompiler } from './';
 import { formatStatsMessages } from './format.js';
 
-/**
- * Add node polyfill tip when failed to resolve node built-in modules.
- */
-const hintNodePolyfill = (message: string): string => {
-  const getTips = (moduleName: string) => {
-    const tips = [
-      `Tip: "${moduleName}" is a built-in Node.js module. It cannot be imported in client-side code.`,
-      `Check if you need to import Node.js module. If needed, you can use ${color.cyan('@rsbuild/plugin-node-polyfill')}.`,
-    ];
-
-    return `${message}\n\n${color.yellow(tips.join('\n'))}`;
-  };
-
-  const isNodeProtocolError = message.includes(
-    'need an additional plugin to handle "node:" URIs',
-  );
-  if (isNodeProtocolError) {
-    return getTips('node:*');
-  }
-
-  if (!message.includes(`Can't resolve`)) {
-    return message;
-  }
-
-  const matchArray = message.match(/Can't resolve '(\w+)'/);
-  if (!matchArray) {
-    return message;
-  }
-
-  const moduleName = matchArray[1];
-  const nodeModules = [
-    'assert',
-    'buffer',
-    'child_process',
-    'cluster',
-    'console',
-    'constants',
-    'crypto',
-    'dgram',
-    'dns',
-    'domain',
-    'events',
-    'fs',
-    'http',
-    'https',
-    'module',
-    'net',
-    'os',
-    'path',
-    'punycode',
-    'process',
-    'querystring',
-    'readline',
-    'repl',
-    'stream',
-    '_stream_duplex',
-    '_stream_passthrough',
-    '_stream_readable',
-    '_stream_transform',
-    '_stream_writable',
-    'string_decoder',
-    'sys',
-    'timers',
-    'tls',
-    'tty',
-    'url',
-    'util',
-    'vm',
-    'zlib',
-  ];
-
-  if (moduleName && nodeModules.includes(moduleName)) {
-    return getTips(moduleName);
-  }
-
-  return message;
-};
-
 function formatErrorMessage(errors: string[]) {
-  const messages = errors.map((error) => hintNodePolyfill(error));
-  const text = `${messages.join('\n\n')}\n`;
   const title = color.bold(color.red('Compile error: '));
 
   if (!errors.length) {
@@ -95,6 +15,7 @@ function formatErrorMessage(errors: string[]) {
   const tip = color.yellow(
     'Failed to compile, check the errors for troubleshooting.',
   );
+  const text = `${errors.join('\n\n')}\n`;
 
   return `${title}\n${tip}\n${text}`;
 }
